pyproject.toml
[tool.poetry]
name = "python-examples"
version = "0.1.0"
description = "examples of common python libs"
authors = ["James Campbell"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.3"
pysocks = "^1.7.1"
hug = "^2.6.1"
rethinkdb = "^2.4.10.post1"
quandl = "^3.7.0"
nltk = "^3.8.1"
exifread = "^3.0.0"
blockchain = "^1.4.4"
websockify = "^0.11.0"
shodan = "^1.31.0"
urllib3 = "^2.2.1"
argparse = "^1.4.0"
fuzzywuzzy = "^0.18.0"
scrapy = "^2.11.1"
pytest = "^8.1.1"
termcolor = "^2.4.0"
pycld2 = "^0.41"
polyglot = "^16.7.4"
tika = "^2.6.0"
pyzillow = "^0.7.0"
geotext = "^0.4.0"
tabulate = "^0.9.0"
tqdm = "^4.66.2"
redis = "^5.0.3"
psycopg2-binary = "^2.9.9"
pypdf2 = "^3.0.1"
pinboard = "^2.1.9"
webdriver-manager = "^4.0.1"
scapy = "^2.5.0"
matplotlib = "^3.8.4"
iptcinfo3 = "^2.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
